refactor: move gesture handler options and logic to a dedicated file
