+++ lawnchair/src/app/lawnchair/gestures/ui/CreateActionsScreen.kt
@@ -1,6 +1,6 @@
 package app.lawnchair.gestures.ui
 
-import android.app.Activity
+import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
@@ -8,14 +8,14 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import app.lawnchair.gestures.config.GestureHandlerConfig
 import app.lawnchair.gestures.config.GestureHandlerOption
+import app.lawnchair.gestures.config.buildConfigFrom
+import app.lawnchair.gestures.config.filterGestureHandlerOptions
 import app.lawnchair.preferences2.preferenceManager2
 import app.lawnchair.ui.preferences.components.controls.ClickablePreference
-import app.lawnchair.ui.preferences.components.gestureHandlerOptions
 import app.lawnchair.ui.preferences.components.layout.PreferenceLayoutLazyColumn
 import app.lawnchair.ui.preferences.components.layout.preferenceGroupItems
 import com.android.launcher3.R
 import com.patrykmichalik.opto.core.firstBlocking
-import kotlin.collections.contains
 import kotlinx.coroutines.launch
 
 @Composable
@@ -27,17 +27,14 @@ fun CreateActionsScreen(
     val scope = rememberCoroutineScope()
 
     val prefs2 = preferenceManager2()
-    val newOptions = gestureHandlerOptions.filterNot { option ->
-        option in listOf(
-            GestureHandlerOption.OpenAppDrawer,
-            GestureHandlerOption.OpenAppSearch,
-        ) &&
-            prefs2.deckLayout.firstBlocking()
-    }
+    val newOptions =
+        filterGestureHandlerOptions(deckLayoutEnabled = prefs2.deckLayout.firstBlocking())
 
     fun onClick(option: GestureHandlerOption) {
         scope.launch {
-            val config = option.buildConfig(context as Activity) ?: return@launch
+            val config = option.buildConfigFrom(context) ?: return@launch
+
+            Log.d("CreateActionsScreen", "config: $config")
             onSelect(config)
         }
     }
@@ -46,7 +43,7 @@ fun CreateActionsScreen(
         label = stringResource(id = R.string.lawnchair_actions),
         modifier = modifier,
     ) {
-        preferenceGroupItems(items = newOptions, isFirstChild = true) { index, it ->
+        preferenceGroupItems(items = newOptions, isFirstChild = true) { _, it ->
             ClickablePreference(
                 label = it.getLabel(context),
                 onClick = { onClick(it) },
